Obsolete the host's Roles https://github.com/Particular/NServiceBus.Host/issues/127
