+++ src/NServiceBus.Hosting.Windows/NServiceBus.Host.csproj
@@ -129,8 +129,7 @@
     <Compile Include="Profiles\Handlers\IntegrationProfileHandler.cs" />
     <Compile Include="Profiles\Handlers\LiteProfileHandler.cs" />
     <Compile Include="Profiles\Handlers\PerformanceCountersProfileHandler.cs" />
-    <Compile Include="Roles\RoleManager.cs" />
-    <Compile Include="Roles\UsingTransport.cs" />
+    <Compile Include="UsingTransport.cs" />
     <Compile Include="InstallWindowsHost.cs" />
     <Compile Include="StartableAndStoppable\StartableAndStoppableFeature.cs" />
     <Compile Include="StartableAndStoppable\StartableAndStoppableRunner.cs" />
