+++ src/NServiceBus.Hosting.Windows/NServiceBus.Host32.csproj
@@ -131,13 +131,12 @@
     <Compile Include="Profiles\Handlers\LiteProfileHandler.cs" />
     <Compile Include="Profiles\Handlers\PerformanceCountersProfileHandler.cs" />
     <Compile Include="Resharper.Annotations.g.cs" />
-    <Compile Include="Roles\RoleManager.cs" />
-    <Compile Include="Roles\UsingTransport.cs" />
     <Compile Include="StartableAndStoppable\EndpointStartableAndStoppableExtensions.cs" />
     <Compile Include="StartableAndStoppable\IWantToRunWhenEndpointStartsAndStops.cs" />
     <Compile Include="StartableAndStoppable\StartableAndStoppableFeature.cs" />
     <Compile Include="StartableAndStoppable\StartableAndStoppableRunner.cs" />
     <Compile Include="TaskEx.cs" />
+    <Compile Include="UsingTransport.cs" />
     <Compile Include="WindowsHost.cs" />
     <Compile Include="Profiles\Integration.cs" />
     <Compile Include="Profiles\Lite.cs" />
